requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
nltk
lightgbm
tqdm
regex
torch
transformers
